third_party/WebKit/LayoutTests/accessibility/scroll-to-make-visible-nested-2.html

<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>

<p id="description"></p>

<div id="container" style="width: 100px; overflow: scroll">
  <div style="border: 1px solid #000; width: 1000px; height: 5000px;">5000-pixel box</div>
  <button id="target">Target</button>
  <div style="border: 1px solid #000; width: 1000px; height: 5000px;">5000-pixel box</div>
</div>

<div id="console"></div>

<script>
description("Tests that scrolling to make an element visible works when the inner scroll container doesn't need to be scrolled, but the other one does.");

function runTest() {
    window.container = document.getElementById("container");
    window.target = document.getElementById("target");

    if (window.accessibilityController) {
        target.focus();
        var targetAccessibleObject = accessibilityController.focusedElement;
    }

    // Reset the initial scroll positions (since calling focus() can scroll the page too).
    window.scrollTo(0, 0);
    container.scrollTop = 0;
    shouldBe("window.pageYOffset", "0");
    shouldBe("container.scrollTop", "0");
    shouldBeGreaterThanOrEqual("target.getBoundingClientRect().top", "5000");

    // Scroll to make target visible.
    if (window.accessibilityController)
        targetAccessibleObject.scrollToMakeVisible();

    // Instead of testing the exact scroll offsets of the two containers, just test that
    // the new absolute position of the target is on-screen.
    shouldBeGreaterThanOrEqual("window.innerHeight", "target.getBoundingClientRect().bottom");

    finishJSTest();
}

runTest();

</script>

</body>
</html>
