third_party/WebKit/LayoutTests/accessibility/scroll-to-make-visible-nested-2-expected.txt

Tests that scrolling to make an element visible works when the inner scroll container doesn't need to be scrolled, but the other one does.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


5000-pixel box
Target
5000-pixel box
PASS window.pageYOffset is 0
PASS container.scrollTop is 0
PASS target.getBoundingClientRect().top is >= 5000
PASS window.innerHeight is >= target.getBoundingClientRect().bottom
PASS successfullyParsed is true

TEST COMPLETE
